+++ src/main/java/fr/raksrinana/fallingtree/config/ToolConfiguration.java
@@ -3,9 +3,10 @@ package fr.raksrinana.fallingtree.config;
 import com.google.common.collect.Lists;
 import net.minecraft.item.Item;
 import net.minecraftforge.common.ForgeConfigSpec;
+import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Stream;
+import static fr.raksrinana.fallingtree.FallingTreeUtils.getAsItems;
 
 public class ToolConfiguration{
 	private final ForgeConfigSpec.ConfigValue<List<? extends String>> whitelisted;
@@ -22,12 +23,12 @@ public class ToolConfiguration{
 		damageMultiplicand = builder.comment("Defines the number of times the damage is applied to the tool (ie: if 1 then breaking 5 logs will give 5 damage; if set to 2, breaking 5 logs will give 10 damage). This only applies when the tree is cut when using the mod.").defineInRange("damage_multiplicand", 1, 1, Integer.MAX_VALUE);
 	}
 	
-	public Stream<Item> getBlacklisted(){
-		return blacklisted.get().stream().map(CommonConfig::getItem).filter(Objects::nonNull);
+	public Collection<Item> getBlacklisted(){
+		return getAsItems(blacklisted.get());
 	}
 	
-	public Stream<Item> getWhitelisted(){
-		return whitelisted.get().stream().map(CommonConfig::getItem).filter(Objects::nonNull);
+	public Collection<Item> getWhitelisted(){
+		return getAsItems(whitelisted.get());
 	}
 	
 	public boolean isIgnoreDurabilityLoss(){